+++ src/BrowserHost/Tab/WebView2/WebView2Browser.cs
@@ -149,8 +149,9 @@ public sealed class WebView2Browser : UserControl, ITabWebBrowser, IDisposable
         _core.FaviconChanged += (_, __) => { _favicon = _core.FaviconUri; actionContextBrowser.UpdateTabFavicon(_id, _favicon); };
         // Mimic CefSharp RequestHandler: open new window requests as background tabs instead of OS windows
         _core.NewWindowRequested += Core_NewWindowRequested;
-        _core.PermissionRequested += Core_PermissionRequested;
         _controller!.AcceleratorKeyPressed += Controller_AcceleratorKeyPressed;
+
+        WebViewPermissionHandler.Register(_core);
     }
 
     private void Core_NewWindowRequested(object? sender, CoreWebView2NewWindowRequestedEventArgs e)
@@ -161,31 +162,6 @@ public sealed class WebView2Browser : UserControl, ITabWebBrowser, IDisposable
             PubSub.Publish(new NavigationStartedEvent(uri, UseCurrentTab: false, SaveInHistory: true));
             e.Handled = true; // Prevent external window
         }
-    }
-
-    private void Core_PermissionRequested(object? sender, CoreWebView2PermissionRequestedEventArgs e)
-    {
-        if (e.PermissionKind != CoreWebView2PermissionKind.Notifications)
-            return; // let default behavior handle other permissions
-
-        var deferral = e.GetDeferral();
-        Application.Current.Dispatcher.BeginInvoke(() =>
-        {
-            try
-            {
-                var origin = new Uri(e.Uri).GetLeftPart(UriPartial.Authority);
-                var result = NotificationPermissionDialog.ShowDialog(MainWindow.Instance, origin);
-                e.State = result ? CoreWebView2PermissionState.Allow : CoreWebView2PermissionState.Deny;
-            }
-            catch
-            {
-                e.State = CoreWebView2PermissionState.Default;
-            }
-            finally
-            {
-                deferral.Complete();
-            }
-        });
     }
 
     private void ApplySettings()
